update timeout examples.
