+++ test/test_bo.py
@@ -1,3 +1,6 @@
+import os
+import sys
+sys.path.append(os.getcwd())
 
 import logging
 import numpy as np
